Move `ExceptionMiddleware` to `starlette.middleware.exceptions`
